error in calculation of d/dx(u*df/dx) detected and solved. refreshing documentation

+++ src/SHOCK.c
@@ -747,9 +747,9 @@ void startSimulation(
 			}
 			if(pnt_config->flag_ManufacturedSolution==1)
 			{
-				if(pnt_config->MPI_rank==0){printf("SHOCK: Recent L2-Delta(%s): %.8Le)\n",
+				if(pnt_config->MPI_rank==0){printf("SHOCK: Recent L2-Delta(%s): %.8e)\n",
 						pnt_config->ManufacturedSolution_L2_Delta_name,
-						pnt_config->ManufacturedSolution_L2_Delta);}
+						(double)pnt_config->ManufacturedSolution_L2_Delta);}
 			}
 			if(pnt_config->MPI_rank==0){printf("SHOCK: --------\n");}
 		}
